+++ pygccxml/declarations/mdecl_wrapper.py
@@ -12,6 +12,7 @@ within the code.
 """
 
 import os
+import warnings
 
 
 class call_redirector_t(object):
@@ -107,8 +108,10 @@ class mdecl_wrapper_t(object):
     def __contains__(self, item):
         return item in self.declarations
 
-    def to_list(self):
-        # Not used internally. Used by py++. TODO: Add test or deprecate ?
+    def to_list(self):  # pragma: no cover
+        warnings.warn(
+            "The to_list method is deprecated. Please implement your own "
+            "version of it if you really need it.", DeprecationWarning)
         li = []
         for d in self.declarations:
             li.append(d)